application/peakflow-tasks/pyproject.toml
[project]
name = "peakflow-tasks"
version = "0.1.0"
description = "Distributed task processing for CodeTrekking fitness data pipeline using Celery and RabbitMQ"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "CodeTrekking Team" }
]
keywords = ["fitness", "data-processing", "celery", "rabbitmq", "tasks"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "celery[redis]>=5.3.0",
    "kombu>=5.3.0",
    "pika>=1.3.0",
    "elasticsearch>=8.0.0",
    "python-dateutil>=2.8.0",
    "pytz>=2023.3",
    "tenacity>=8.2.0",
    "structlog>=23.1.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "typing-extensions>=4.5.0",
    "peakflow",
    "pydantic-settings>=2.10.1",
    "flower>=2.0.1",
    "sqlmodel>=0.0.14",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.12.0",
    "matplotlib>=3.10.6",
    "seaborn>=0.13.2",
    "pandas>=2.3.1",
    "numpy>=2.3.2",
    "garmin-fit-sdk>=21.178.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-celery>=0.1.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.280",
    "mypy>=1.5.0",
    "pre-commit>=3.3.0",
    "watchdog>=3.0.0",
    "colorama>=0.4.6",
    "python-dotenv>=1.0.0",
]

test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-celery>=0.1.0",
    "pytest-mock>=3.11.0",
    "pytest-cov>=4.1.0",
]

monitoring = [
    "flower>=2.0.0",
    "prometheus-client>=0.17.0",
]

[project.scripts]
peakflow-tasks = "peakflow_tasks.cli:cli"
peakflow-worker = "peakflow_tasks.cli:worker"
peakflow-beat = "peakflow_tasks.cli:beat"
peakflow-flower = "peakflow_tasks.cli:flower"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["peakflow_tasks"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=peakflow_tasks",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "-v"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "load: Load and performance tests",
    "slow: Slow tests that take more than 30 seconds",
]

[tool.coverage.run]
source = ["peakflow_tasks"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.uv.sources]
peakflow = { path = "../peakflow", editable = true }

[dependency-groups]
dev = [
    "colorama>=0.4.6",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "python-dotenv>=1.1.1",
    "watchdog>=6.0.0",
]
